requirements.txt
faiss-cpu
sentence-transformers
openai
requests
pyyaml
